Don't restart if replaying a demo
If you get a checkpoint and very quickly restart, it might be possible for the
replay to bug out as the restart can go through before the checkpoint
has been collected. I wasn't able to get it to happen normally, using
(the best I could get of) frame perfect restarts across high and low FPS
values, but I WAS able to get it to occur by recording a level with a
checkpoint and then deleting the checkpoint from the level.