A few minor additions
Versions of Pianoforte from 1.50 and 1.14 added to the jukebox, similar to the legacy soundtracks already in-game.

Added an option to disable taunts in Multiplayer.

Added "waiting.png" from MBP 1.50 to the MBP UI texture pack.

Debug mode will no longer spam the console with unused mode callback messages.

You are no longer able to select Pianoforte the be the song for a level, since it wouldn't do anything.