Update editorgui.cs
Taking icon and preview pictures now returns to the previous GUI afterwards rather than always going back to the level editor
In normal play, this wouldn't matter, but this allows using console commands for this as well without opening the editor first. Otherwise, the game crashes because the editor gui hasn't been loaded yet. Plus, even if it was loaded, being sent to the editor afterwards if you weren't in there before would be weird anyway.