Play Button Replacements (Border Color)
Just committing this in case this wants to be looked at again, but border color isn't nearly as versatile as I thought it was. But just replacing the texture entirely with a solid color is super uggo.

Currently, the play button changes to a "Set Finale" icon when an RTA speedrun has been started without a set destination, and a "Start Run" icon when an RTA speedrun with a set destination is waiting to start.