Fixed PhysMod Trigger Position
Silly me! I assumed the water physics trigger present in this area was actually in the same place as all those water planes. Therefore, I thought it'd be simple enough to just give the new physmod trigger the same position and dimensions as that! What a fool I was...

Seriously, this level has a water physics trigger. But it's _behind_ the water wall. Where you can never get to. Incredible work, Uil.